+++ resources/application.go
@@ -4,7 +4,6 @@ import (
 	"errors"
 
 	"github.com/generationtux/brizo/database"
-	"github.com/generationtux/brizo/kube"
 	"github.com/jinzhu/gorm"
 	"github.com/pborman/uuid"
 )
@@ -12,9 +11,9 @@ import (
 // Application as defined by Brizo.
 type Application struct {
 	database.Model
-	UUID string     `gorm:"not null;unique_index" sql:"type:varchar(36)"`
-	Name string     `gorm:"not null;unique_index"`
-	Pods []kube.Pod `gorm:"-"` // gorm will ignore, but we can populate
+	UUID string `gorm:"not null;unique_index" sql:"type:varchar(36)"`
+	Name string `gorm:"not null;unique_index"`
+	Pods []Pod  `gorm:"-"` // gorm will ignore, but we can populate
 }
 
 // BeforeCreate is a hook that runs before inserting a new record into the database
@@ -48,14 +47,15 @@ func UpdateApplication(db *gorm.DB, app *Application) (bool, error) {
 }
 
 // GetApplication will get an existing Application by name
-func GetApplication(db *gorm.DB, id string, getPods kube.PodRetrieval) (*Application, error) {
+func GetApplication(db *gorm.DB, id string, getPods PodRetrieval) (*Application, error) {
 	if _, valid := uuid.Parse(id).Version(); valid {
 		app := new(Application)
 		if err := db.Where("uuid = ?", id).First(&app).Error; err != nil {
 			return app, err
 		}
 
-		app.Pods, err = getPods(app)
+		pods, err := getPods(app.UUID)
+		app.Pods = pods
 
 		return app, err
 	}
